backend/requirements.txt
Django==4.0.2
asgiref==3.5.0
django-cors-headers==3.11.0
django-kronos==1.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
